Moved WidgetUpdater to be an inner interface 'Updater' of 'Widget' class. Renamed Compose widget to StatelessCompose
