+++ samples/src/main/java/utils/MainMenuButton.java
@@ -6,13 +6,11 @@ import dev.gamekit.ui.events.MouseEvent;
 import dev.gamekit.ui.widgets.*;
 
 @CustomWidgetBuilder
-public class MainMenuButton extends Compose {
+public class MainMenuButton extends StatelessCompose {
   @CustomWidgetBuilderField
   public String text;
   @CustomWidgetBuilderField(comparable = false)
   public MouseEvent.Handler mouseListener;
-
-  private int clickCount = 0;
 
   public MainMenuButton(MainMenuButtonConfig config) {
     super(config);
@@ -36,19 +34,12 @@ public class MainMenuButton extends Compose {
   @Override
   protected Widget build() {
     return Button.create(
-      props -> props.mouseListener = (ev) -> {
-        if (ev.type == MouseEvent.Type.CLICK) {
-          clickCount++;
-          updateUI();
-        }
-
-        mouseListener.handleEvent(ev);
-      },
+      props -> props.mouseListener = mouseListener,
       Padding.create(
         12, 12, 16, 12,
         Text.create(
           props -> {
-            props.text = text + " " + clickCount;
+            props.text = text;
             props.fontSize = 20;
             props.fontStyle = Text.BOLD;
           }
